src/python/processing_function/requirements.txt
azure-functions==1.0.0b5
ai_acc_quality
pypyodbc==1.3.5.2
azure-storage >= 0.34.3
